TableDescriptor.drop_descriptor_table now only drops the table when it exists

+++ lib/cassandra_model/table_descriptor.rb
@@ -15,7 +15,7 @@ module CassandraModel
       end
 
       def drop_descriptor_table
-        session.execute("DROP TABLE #{table_name}")
+        session.execute("DROP TABLE #{table_name}") if descriptor_table_exists?
       end
 
       private
